refactor(team): extract shared TeamPending and InviteTeamAction
