+++ src/routes/_authedRoutes/profile/barbershops/team/invitations/index.tsx
@@ -1,7 +1,5 @@
 /** biome-ignore-all lint/style/noNonNullAssertion: barbershop is primed by the loader and gated to owners/staff */
 
-import type { Barbershop } from "@convex/schema";
-import { UserPlusIcon } from "@phosphor-icons/react";
 import { createFileRoute, redirect } from "@tanstack/react-router";
 import { lazy, Suspense } from "react";
 
@@ -13,15 +11,17 @@ import {
   DashboardPageHeading,
   DashboardPageStats,
 } from "@/components/dashboard/dashboard-page";
+import {
+  InviteTeamAction,
+  TeamPending,
+} from "@/components/dashboard/team-page-shared";
 import { ProfileTabSkeleton } from "@/components/layout/skeleton/profile-tab-skeleton";
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardDescription,
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
 import { cacheTime } from "@/config/cache";
 import { useBarbershopByMemberUserId } from "@/hooks/barbershop/use-barbershop";
 import { useBarbershopMemberRoles } from "@/hooks/barbershop/use-barbershop-member";
@@ -30,12 +30,6 @@ import {
   useInvitations,
 } from "@/hooks/use-invitations";
 import { useSession } from "@/hooks/use-session";
-
-const InviteBarberDialog = lazy(() =>
-  import("@/components/barbers/invite-barber-dialog").then((module) => ({
-    default: module.InviteBarberDialog,
-  })),
-);
 
 const InvitationsList = lazy(() =>
   import("@/components/barbers/invitations-list").then((module) => ({
@@ -67,53 +61,6 @@ export const Route = createFileRoute(
     }
   },
 });
-
-function TeamPending() {
-  return (
-    <DashboardPage>
-      <DashboardPageHeader>
-        <div className="space-y-2">
-          <Skeleton className="h-6 w-32" />
-          <Skeleton className="h-4 w-72 max-w-full" />
-        </div>
-        <Skeleton className="h-9 w-24" />
-      </DashboardPageHeader>
-      <DashboardPageContent>
-        <ProfileTabSkeleton />
-      </DashboardPageContent>
-    </DashboardPage>
-  );
-}
-
-function InviteTeamAction({
-  barbershopId,
-  canInviteStaff,
-}: {
-  barbershopId: Barbershop["_id"];
-  canInviteStaff: boolean;
-}) {
-  return (
-    <Suspense
-      fallback={
-        <Button disabled>
-          <UserPlusIcon />
-          Invitar
-        </Button>
-      }
-    >
-      <InviteBarberDialog
-        barbershopId={barbershopId}
-        canInviteStaff={canInviteStaff}
-        trigger={
-          <Button>
-            <UserPlusIcon />
-            Invitar
-          </Button>
-        }
-      />
-    </Suspense>
-  );
-}
 
 function RouteComponent() {
   const { data: user } = useSession();
